Fix #13
As this is a fairly major change in the protocol, and because it changes
the syntax of the PING command I've renamed PING to HELLO (which is a
bit more descriptive of its function given that it is now required to
precede any commands from a client).

Also refactored exceptions, warnings, and network repeater codes into
their own modules.