Simplify protocol maintenance
In an effort to simplify protocol maintenance, use an abstract class to
define the protocol, its commands, their parameters, types, defaults,
etc. We'll then use decorators or functions to convert the abstract
class for other uses (to parse incoming commands to the server, to
format commands for the client, and to document the protocol itself).
Documentation generation is now done.