Work on #26
This is more or less done as far as the batch client goes (although we
still need to fix setting the entire list of servers manually). I've
added a "sort" command to cpi which is all that's really practical there
(can't imagine manually specifying the order of a pile of servers by
typing...). The cpigui client needs lots of work though (especially to
incorporate recording stuff too).