Add close method to client
Added close() method to the CompoundPiClient class and the ability to
use it as a context manager which automatically calls close(). Re-wrote
most of the documentation to use the context manager style. This ensures
that the listening server for download connections is properly closed
down at the end of the script. The command line and GUI clients are
updated accordingly.